Update StockPricePrediction.py
I tried the initial code, but there were some problems in the CSV file. I tried with another CSV file from yahoo. I also found an anomaly that the ohlc_avg was in millions, so I removed 'volume' column from the table for making the average. And I did remove 'dataset = dataset.reindex(index = dataset.index[::-1])' because due to it, the price was falling, whereas it should have increased with time.

I tried the code in this notebook, you can review it too.
https://colab.research.google.com/drive/1v6FF9zHMYgBy8QD1kuSL3q3ln-pD7lQw?usp=sharing

And thanks to you for sharing this code, so that I could learn LSTM. 
You can contact me for any clarifications at samar.080301@gmai.com or +918617781030(WhatsApp & Call)